Update to single API Test Table
API Test Table now is a single table with all APIs. All inputs are text inputs and are only enabled if needed (e.g. input 1 and 2 are enabled for only certain APIs).

Also made the API table show up be default and allow the user to change to getFillerChars if desired.