Telugu characters in response and fixes
Adjusted it so that all APIs that have the potential to return Telugu, do so in their full form, versus unicode points.

Fixed split word bug and updated tests to properly fetch and display arrays and objects.

Changed the parser to only process the input when a button is pressed, versus automatically.